# estudio_madame/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Estúdio Madame API
  description: API para gerenciamento de galerias fotográficas com integração Google Drive
  version: 1.0.0
  contact:
    name: API Support
    email: support@estudiomadame.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.estudiomadame.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token para autenticação
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          format: uri
          nullable: true
        galleries_count:
          type: integer
        last_activity:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateClientRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
        avatar:
          type: string
          format: uri

    # ==================== GALLERY SCHEMAS ====================
    GallerySettings:
      type: object
      properties:
        privacy:
          type: string
          enum: [public, private, password]
        password:
          type: string
          nullable: true
        allow_downloads:
          type: boolean
        mature_content:
          type: boolean
        max_client_selection:
          type: integer
          nullable: true
        layout:
          type: string
          enum: [grid, masonry, list]
        default_sort:
          type: string
          enum: [newest, oldest, az]
    GalleryStats:
      type: object
      properties:
        views:
          type: integer
        downloads:
          type: integer
        selections:
          type: integer
    Gallery:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        client_id:
          type: string
          format: uuid
        client_name:
          type: string
        status:
          type: string
          enum: [draft, published, client_selection, archived]
        cover_image:
          type: string
          format: uri
          nullable: true
        photo_count:
          type: integer
        date_created:
          type: string
          format: date-time
        event_date:
          type: string
          format: date
          nullable: true
        location:
          type: string
          nullable: true
        access_token:
          type: string
          nullable: true
        google_drive_folder_id:
          type: string
          nullable: true
        google_drive_folder_name:
          type: string
          nullable: true
        auto_sync_enabled:
          type: boolean
        last_sync_at:
          type: string
          format: date-time
          nullable: true
        sync_status:
          type: string
          enum: [idle, syncing, error]
        settings:
          $ref: '#/components/schemas/GallerySettings'
        stats:
          $ref: '#/components/schemas/GalleryStats'
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
paths:
...
  /clients:
    get:
      tags:
        - Clients
      summary: Listar todos os clientes
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          description: Buscar por nome ou email
          schema:
            type: string
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Client'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - Clients
      summary: Criar novo cliente
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Cliente criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Client'
        '400':
          description: Dados inválidos
  /clients/{id}:
    get:
      tags:
        - Clients
      summary: Obter detalhes de um cliente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhes do cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Client'
        '404':
          description: Cliente não encontrado

    patch:
      tags:
        - Clients
      summary: Atualizar cliente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                phone:
                  type: string
                avatar:
                  type: string
      responses:
        '200':
          description: Cliente atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Client'

    delete:
      tags:
        - Clients
      summary: Deletar cliente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cliente deletado com sucesso
        '404':
          description: Cliente não encontrado
  /clients/{id}/galleries:
    get:
      tags:
        - Clients
      summary: Listar galerias de um cliente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de galerias do cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Gallery'

  # ==================== APPROVALS ====================
...